+++ Civi/ActionProvider/Action/Event/CreateOrUpdateEvent.php
@@ -58,6 +58,8 @@ class CreateOrUpdateEvent extends AbstractAction {
       new Specification('end_date', 'Timestamp', E::ts('End date'), false, null, null, null, FALSE),
       new Specification('is_active', 'Boolean', E::ts('Is active'), false, 1, null, null, FALSE),
       new Specification('is_public', 'Boolean', E::ts('Is public'), false, 0, null, null, FALSE),
+      new Specification('max_participants', 'Integer', E::ts('Max. Participants'), false, null),
+      new Specification('waitlist_text', 'String', E::ts('Text when event is full'), false, null),
     ));
 
     $config = ConfigContainer::getInstance();
@@ -145,6 +147,12 @@ class CreateOrUpdateEvent extends AbstractAction {
     }
     if ($parameters->doesParameterExists('is_public')) {
       $apiParams['is_public'] = $parameters->getParameter('is_public');
+    }
+    if ($parameters->doesParameterExists('max_participants')) {
+      $apiParams['max_participants'] = $parameters->getParameter('max_participants');
+    }
+    if ($parameters->doesParameterExists('waitlist_text')) {
+      $apiParams['waitlist_text'] = $parameters->getParameter('waitlist_text');
     }
 
     foreach($this->getParameterSpecification() as $spec) {